fix: appEngineDeployAndCleanUp() and appEnginePublish() should succeed on first attempt
Close #42

Log analysis revealed that the failure to deploy was reported as an access error
to the deployment storage bucket (after all files had been uploaded to the storage
bucket). Some experimentation suggested that it was basically a timing issue. At
the moment, pausing to get a description of the app seems to be enough to
trigger/delay whatever needs to happen long enough that the subsequent deploy
works.