feat!: reorganized batch scripts for consistency/DRY-ness
BREAKING: The logic of the `env` argument to appEnginePublish() and
appEngineDeployandCleanup() has been modified: it now expects either
`true` (read values from the default environment) or a path to a
`.env` file from which to read the values. 1Password support is now
incorporated, encouraging us all to store our secrets more securely.